Added actions and reactive model.
Added a concept of actions to make it easier to observe to things
happening on the view. Updated the model so that it doesn’t update the
view through a consumer but rather allows the intent to subscribe to
changes to itself (the model) and when the model changes the intent can
determine how to change things.